% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-site.R
\name{format.cacs_site_plot_pipeline}
\alias{format.cacs_site_plot_pipeline}
\title{Describe the four maps in one line}
\usage{
\method{format}{cacs_site_plot_pipeline}(x, ...)
}
\arguments{
\item{x}{A list of class \code{"cacs_site_plot_pipeline"}, as returned by
\code{\link[=cacs_plot_site_pipeline]{cacs_plot_site_pipeline()}}.}

\item{...}{Not used.}
}
\value{
A string.
}
\description{
Returns a one-line description of the list that names its class and the
\code{site_id} of its site, such as
\code{"<cacs_site_plot_pipeline: 4 widgets for site AL_BHM_01>"}, for use in
text (for example, \code{sprintf("Maps: \%s", format(x))}).
}
\examples{
if (requireNamespace("leaflet", quietly = TRUE)) {
  fx <- readRDS(system.file("extdata", "visual_walkthrough_fixture.rds",
                            package = "catchmentACS"))
  maps <- cacs_plot_site_pipeline(
    site_id = "AL_BHM_01", iso_sf = fx$iso_sf, tract_sf = fx$tract_sf,
    acs_sf = fx$acs_sf, run_result = fx$run_result, sites_df = fx$sites_df
  )
  sprintf("Maps: \%s", format(maps))
}
}
\seealso{
Other maps of one site:
\code{\link[=cacs_plot_site_intersection]{cacs_plot_site_intersection()}},
\code{\link[=cacs_plot_site_isochrone]{cacs_plot_site_isochrone()}},
\code{\link[=cacs_plot_site_pipeline]{cacs_plot_site_pipeline()}},
\code{\link[=cacs_plot_site_rates]{cacs_plot_site_rates()}},
\code{\link[=cacs_plot_site_weighted]{cacs_plot_site_weighted()}},
\code{\link[=print.cacs_site_plot_pipeline]{print.cacs_site_plot_pipeline()}}
}
\concept{maps of one site}
